+++ wp-property.php
@@ -39,39 +39,10 @@ if (!defined('WPP_Templates')) {
   define('WPP_Templates', WPP_Path . 'static/views');
 }
 
-if (!function_exists('parse_feature_flags')) {
-  /**
-   * Set Feature Flag constants by parsing composer.json
-   *
-   * @todo Make sure settings from DB can override these.
-   *
-   * @author potanin@UD
-   * @return array|mixed|null|object
-   */
-  function parse_feature_flags()
-  {
-    try {
-      $_raw = file_get_contents(plugin_dir_path(__FILE__) . 'composer.json');
-      $_parsed = json_decode($_raw);
-      // @todo Catch poorly formatted JSON.
-      if (!is_object($_parsed)) {
-        // throw new Error( "unable to parse."  );
-      }
-      foreach ((array)$_parsed->extra->featureFlags as $_feature) {
-        if (!defined($_feature->constant)) {
-          define($_feature->constant, $_feature->enabled);
-        }
-      }
-    } catch (Exception $e) {
-      echo 'Caught exception: ', $e->getMessage(), "\n";
-    }
-    return isset($_parsed) ? $_parsed : null;
-  }
-
-  // Init feature flags
-  parse_feature_flags();
-
-}
+/**
+ * Set Feature Flags
+ */
+include_once WPP_Path . 'feature-flags.php';
 
 // Use Freemius is flag is enabled.
 if (defined('WPP_FEATURE_FLAG_FREEMIUS') && WPP_FEATURE_FLAG_FREEMIUS) {